api favourites

+++ app/Http/Controllers/Api/Auth/FavouriteController.php
@@ -8,6 +8,7 @@ use App\Http\Resources\BaseCollection;
 use App\Http\Resources\OrderResource;
 use App\Models\Advertisement;
 use App\Models\Favourite;
+use App\Models\Order;
 use Illuminate\Http\Request;
 
 class FavouriteController extends Controller
@@ -42,14 +43,22 @@ class FavouriteController extends Controller
         ]);
         $validator['model_type'] = "App\\Models\\$request->model_type";
         $validator['user_id'] = auth()->user()->id;
-//dd($validator['model_type']);
 
-        $fav = Favourite::whereModelId($validator['model_id'])->whereModelType($validator['model_type'])->whereUserId(auth('api')->user()->id);
-        if ($fav->exists()) {
-            $fav->delete();
-        } else {
-            Favourite::create($validator);
+        $order = Order::whereId($validator['model_id'])->exists();
+        $adv = Advertisement::whereId($validator['model_id'])->exists();
+
+        if ($order || $adv){
+            $fav = Favourite::whereModelId($validator['model_id'])->whereModelType($validator['model_type'])->whereUserId(auth('api')->user()->id);
+            if ($fav->exists()) {
+                $fav->delete();
+            } else {
+                Favourite::create($validator);
+            }
+        }else{
+            return \responder::error('wrong model id');
         }
+
+
         return \responder::success('Success');
     }
